Contrast-Stretch Dialog close-button signals "accepted"
Re-enabled the "close" button on the contrast-stretch dialog, since on
small screens the "OK" and "Cancel" buttons are too far below the bottom
of the screen to reach.

The "close" button was originally disabled because we didn't know
whether closing would mean "OK" or "Cancel".  But, since the dialog's
effects are applied immediately, the dialog now interprets "close" as
"accepted."  (This is different from the default Qt behavior of
interpreting it as "rejected.")