Improve usability of spectrum info editor
Spectra collected from raster files now have a flag that controls
whether the name is user-specified or auto-generated.  This flag is
exposed in the user interface, and allows the application to be more
intelligent about name management.

The spectrum info editor dialog is also much more intelligent now.
When name auto-generation is on, changing the configuration of the
spectrum will regenerate the name as expected.  When the user starts
to edit the spectrum name, auto-generation is turned off automatically.

A few other tweaks were made to the dialog code to ensure that spectra
were not changed if the Cancel button is pressed.