New and improved application configuration support
WISER now has a configuration dialog that exposes many of the general
configuration options, including whether the app should send bug reports
to BugSnag.

Configuration is managed in a new ApplicationConfig class, which is held
by the ApplicationState class.  This class is manipulated by the above
dialog.

WISER also persists the configuration state to disk, so that the user
doesn't have to enter it every time.  The location this file is saved
to is OS specific, conforming to the typical practices for Windows,
macOS and Linux/UNIX.

Many other components that use this configuration now access the config
through the application state; hence the large set of changes.  Colors
for pixel cursors, viewport highlights, and other details like the
wavelength for red/green/blue light are now pulled from configuration
to make the application much more flexible.

This commit also includes a bug fix for discarding all collected
spectra; previously this did not work correctly (only the first
collected spectrum was discarded).