Bug-reporting dialog on first run; refactoring
When WISER runs for the very first time, and it creates its config
files, it will now ask the user to opt in for online bug reporting.

Refactored the bug-reporting code into a separate module to make the
operations simpler and more self-contained.

For testing, added a new "generate an error" button that can be turned
on with a feature-flag, so that the bug-reporting functionality can be
exercised.