First cut at bandmath-evaluation support
This commit implements bandmath evaluation including addition,
subtraction, multiplication and division.  Values can be numbers,
images, image bands and spectra.  The operations follow the
researchers' specifications for how these things should work.

A new feature was added to get the entire NumPy array for a raster image
dataset, since some operations must work against the entire image.

This code is untested, and also, even if it is correct it is likely to
be very slow.  Subsequent work will be needed to identify whether such
issues exist and how to resolve them.