Mostly-functional band-math dialog
The band-math dialog is much more complete now:

The dialog now lets the user specify the type of each variable.  Based
on the type, the user can choose different items that the app knows
about, whether they be datasets, datasets+bands, or spectra.

The dialog also guesses a variable's type based on the name.  Normally,
people choose names like "i*" for images, "s*" for spectra, and "b*" for
bands.  So, this isn't hard to do at all.  In general we default to a
variable being a band, since it is called "band-math" after all.

The dialog is able to generate a dict of variable bindings, ready to
pass into the band-math evaluator.

The raster-pane that exposes the band-math dialog is able to invoke the
evaluator with the results of the dialog, but the results are not yet
displayed.