Load dataset from numpy array; display bandmath image or band
Continuing the work on bandmath support.  Since the bandmath evaluator
is producing images and bands, I went ahead and wrote a way to turn a
NumPy array back into a RasterDataSet so that we can try to display
them.  It kind of works.

GDAL provides support for turning NumPy arrays into GDAL Dataset objects
via the gdal_array package.  The code for this is probably at the bare
minimum at the moment; more work will be ongoing.