First cut of WISER plugin support
The doc/plugins.md captures the current plan for where plugins will be
exposed in WISER.  In brief:  Plugins will be supported in the Tools
application menu, in pop-up context menus, and to supply custom
band-math functions.  There are many questions regarding dependencies
that remain to be answered.

The new plugins package provides abstractions for these kinds of
plugins.  There are also two examples:  a Tools-menu plugin (Hello
World!), and a band-math custom function plugin.

The application config is updated to record the plugins being used in
WISER, and these are instantiated when the application starts.  The
application state also holds onto the instantiated plugins, as plugins
that are only referenced from Qt menu actions get garbage-collected,
sigh.

To support custom band-math functions, a few additional details needed
to be exposed in the band-math abstractions.  This will undoubtedly
continue to be refined.