More plugin config UI implementation
Hooked up the "Add" and "Delete" plugin-list buttons.

Added a "Verify" button to the plugin-list section, to verify that all
specified plugin classes are able to be instantiated, and that each
instantiable object is of an expected type.  The system path is updated
before verification occurs, and it is restored after verification, in
case the user needs to make alterations after verification.