Band-math refactoring and UI improvements
The band-math functionality is taking a large step forward with this
commit.  One necessary feature for the band-math UI is the ability to
analyze the expression to see what result will be created, and also
how large it may be, so that users can be warned if the result will
be very large and costly to compute.  This commit includes a first
step in that direction.

Functions and operators have a more advanced abstraction now, allowing
them to report their return-type based on argument types, as well as
some documentation information.  Besides supporting the above
functionality, this also allows user-provided band-math functions to
provide documentation where necessary.

The band-math UI now has a passable workflow for saved expressions,
which can be loaded from a file and/or saved to a file.  I think the
workflow is not overcomplicated, and does the job adequately.  There
are still a few wrinkles to iron out, such as the fact that expressions
can't be deleted from the saved-expressions list.