Enhance logging support within WISER
With the addition of WISER plugin support, logging needs to be enhanced
to make it easier to debug plugins.

WISER now configures logging programmatically using
logging.config.dictConfig() so that it doesn't need to somehow find the
logging config-file.  We may eventually want to go back to a file-based
configuration so that users can tweak the WISER config output, but we
can cross that bridge in the future.

Config and initialization operations now use logging for informational,
debugging and error output.