Refactor raster utility functions into separate module
In preparation for releasing the plugin API, the utility operations
in some modules are being grouped into a single module.  This will
allow the types to be published in documentation, but keep the
utility operations internal-only.

This commit also includes some test refactoring, and some new tests
for various tasks.