Remove warning about "virtual memory read failure"
In general we prefer to read using virtual-memory mapping, to minimize
the amount of memory required to load large datasets.  The GDAL loader
prints a warning when virtual memory cannot be used.  However, this
situation arises more often than expected, particularly in the context
of band-math expressions, which often work entirely in memory.  So, I
am removing the warning here because it doesn't seem to be so relevant.