Refactor code for spectra and spectral calculations
Condensed the spectral code down to a single Python module spectrum.py
(except for the spectral export code).  This will help simplify some
of the imports, since there wasn't a good reason why the code was in
multiple files.

Renamed the SpectrumInfo base-type to Spectrum, and the spectrum_info
module to spectrum.  This helps the code be cleaner and a bit more
streamlined.

This commit also includes some tweaks to the logging output for
band-math calculations.