Refine and document context-menu plugin details
This commit includes a first-draft of the context-menu plugin
documentation, complete with all details of what plugins receive from
WISER when they are invoked.  The documentation includes recommended
usage patterns to play well with WISER.

The context-menu plugin API and implementation are also refined and
refactored in a few ways:

*   The raster-display context-menu invocations now receive more
    information from WISER, such as the current display bands.

*   The code that calls the context-menu plugins uses Python
    keyword-arguments to specify the context values, making the WISER
    internal implementation look a bit cleaner.

*   The context now includes a reference to a WISERControl object that
    will eventually allow plugins to access and control parts of WISER
    that ought to be exposed.  Right now this WISERControl object is
    empty; we will need to grow out the API.

Now that a WISERControl type exists, this is also passed to the
Tools-menu plugins, so that they may also access and control WISER
state.