Add features to WISERControl, RasterDataSet classes
Added functionality to WISERControl so that plugins can make datasets
from numpy arrays, and add datasets to what WISER is displaying.  This
was pleasantly straightforward.

Started some updates to the RasterDataSet and GDALRasterDataSet classes
to handle metadata better - then realized that the metadata support will
really need to be revised to support mutable datasets.  So, that will be
in future work.

Since it's very difficult to mutate a dataset in place, I updated the
plugin documentation so that the example doesn't include this kind of
approach.  Now it always generates a new image after the Gaussian blur
has been applied.