Fix unnecessary stretch-calculation errors
Stretch calculations map image data to the range [0, 1].  Some of the
calculations were going _just over_ 1.0 (probably by the minimum epsilon
that floats can represent), but this would cause the code to raise
exceptions about it.  So, the code now checks that values are in the
range [0, 1 + EPSILON], where EPSILON = 1e-6, and then clamps things to
the range [0, 1].  This should prevent the unnecessarily verbose
complaints due to floating point vagaries.