Improve band-math analysis; refactoring
Band-math analysis can now identify more useful information about
band-math expressions, including the estimated size of the result in
bytes.  This is helpful for giving users feedback on the complexity
of their expressions.

Refactored and renamed some of the files and classes to make the code
more readable.  Also factored some common operations into helper
functions to make the implementation cleaner and shorter.