Add colormap option to the band-chooser UI
In grayscale mode, the band-chooser allows the user to choose whether
or not to apply a color-map to the grayscale data, and if so, what
color-map to use.  The color-map list is populated from matplotlib.
The selected color-map is also displayed in the dialog so the user
knows what they are selecting.