Incorporate color-map into grayscale image display
Refactored the rasterview code that generates a display image from the
raster data set, so that RGB and grayscale are on different paths.

Updated the grayscale rendering to take a colormap, and use it to map
the gray-values to RGB colors.  NumPy array indexing is used to make
this super duper fast.