Improved band-math capabilities and anlysis performance
Improved the band-math capabilities with these new features:

*   Functions can now specify named arguments (like Python functions can
    specify named arguments), for specifying configuration and other
    optional values that can modify a function's behavior.
*   Strings (either single-quoted or double-quoted) can be specified as
    literal values.  Strings cannot be used with operators; they are
    only for mode arguments and the like.
*   The power and unary-negate operations are fully implemented.
*   The band-math dialog reports useful errors when the band-math
    expression evaluates to something other than an image, a band, or
    a spectrum.

The band-math analysis performance is also greatly improved.  The main
issue previously was that the object (image, band or spectrum) needed to
be converted into a NumPy array to get its dimensions and element-type,
and this made analysis very slow.  Now this information is retrieved
from the metadata of the object, which is much much faster.