Refactoring band-math checks; fix save/load band-math bugs
Refactored a lot of the code that checks for band-math issues to reduce
code duplication.  Also refactored the code that converts a
variable-type and shape into a human-readable string so that it can be
used both in error-reporting and in normal analysis code.

Fixed a couple of bugs in the "save band-math expressions" and "load
band-math expressions" code, so it seems to work well now.  Also updated
the "Save expression" code to show the newly-saved expression in the
"saved expressions" combobox when it is saved.