Fix bugs in update_cwd_from_path() method
This method couldn't handle simple filenames as arguments.  This was
further causing an exception because I forgot to import the warnings
module.  The fix was simple - just pass the argument through
os.path.abspath() (prepends the current working directory onto
relative paths and then normalizes the path).  Also, the missing import
was added.