Partial implementation of saving raster data
This is a first attempt to implement saving of raster data in WISER.
The UI has been updated with menu items to save and close data sets,
and the app has a "save as" dialog now.

The main issue is that GDAL doesn't seem to support all the ENVI
header configuration that we need - simple stuff like the "data-ignore
value."  So, not sure if we will be able to continue using GDAL in the
long run.