Generate ENVI .hdr file separately at save
The GDAL ENVI driver doesn't know how to save certain ENVI attributes,
which are pretty important to be able to save.  Two examples are "data
ignore value" and "default bands".  The solution implemented in this
commit is to use GDAL to write out the ENVI image and header files, then
re-generate the header file using WISER's own ENVI file-format support.
The key data-format parameters are copied from the GDAL file.

So far, this approach seems to solve the problem.