Remove WISERControl wrapper; pass ApplicationState directly to plugins
The WISERControl wrapper was a good idea, but required a lot of extra
effort to fill in the implementation.  Since we are basically exposing
the ApplicationState operations directly to the plugins, why not just
go for it and let plugins access and mutate application state through
it.  So that's what this commit does.