Fix usability issue with band-math dialog
It just seems natural to press "Enter" at the end of entering a
band-math expression.  But, the default Qt behavior is for this to close
the band-math dialog.  To fix this, an event-filter was added to the
QLineEdit for the expression, and when Return/Enter is pressed, the
key-press triggers expression analysis instead of closing the dialog.