Fix spectrum-plot click-drag selection issue
If the spectrum-plot window is narrow, picking a point on a spectrum
can cause the plot to resize in very irritating ways.  This becomes
exacerbated when the point is dragged so that the text is updating
a lot.

This is fixed now; the fix was simply to make the text ignored when
matplotlib is laying out the plot.