Spectrum-plot usability improvements
When points are picked on the spectrum-plot display, the data value is
now shown with only the necessary resolution, based on the element-type
of the spectrum being displayed.

When a spectral library is loaded, its spectra will automatically be
displayed if the number of spectra is below some threshold.  This makes
it so that small libraries are easy to see immediately, but large
libraries won't overwhelm WISER.