Initial version of ROI import (GeoJSON)
The ROI import/export code now includes support for importing Regions of
Interest using GDAL/OGR geometries and features.

To support importing ROIs, the ROI/ApplicationState code was changed to
support auto-naming/auto-renaming ROIs to be unique, and to assign
colors to ROIs that don't have colors.

There is still one bug in the ROI code; the UI doesn't get updated when
an ROI is added.