Fix ROI redraw issues
Regions of Interest were not being properly drawn across raster-views
when they were added, edited or deleted, or when selections on ROIs were
added, edited or deleted.  This is now fixed.

Evidently I had thought about this before, because there was a mechanism
for RasterPanes to signal to others that an ROI had been changed.  I
merely completed this mechanism, and have all the ROI-editing
task-delegates emit this signal.  The top-level application receives
this signal and updates the UIs of all other raster-views.