Refactor/restructure spectral library code
The spectral-library code needed to be restructured to support spectra
that may not have the same numbers of bands, or the same other metadata,
for each spectrum in the library.  Things seem to still work after the
update, but it will require more testing.

This change is necessary to support importing spectra from text files,
where each spectrum in the text file may have a different number of
bands and/or other metadata.