More advanced-save functionality; image subsetting works
More advanced-save dialog work.  The "description" value is now saved
to the target file.

The biggest improvement is that a spatial subset of the image is now
saved to the output file when specified by the user.

Also fixed a bug in the "Advanced Save" dialog where the image size
wasn't being displayed properly.