Start adding support for GeoTIFF format to WISER
GeoTIFF format is also supported by GDAL, but we didn't have a
"dataset-implementation" for it to expose its metadata into WISER.

There is now an implementation of this, but right now it defaults to the
general information exposed by GDAL, which clearly doesn't include all
the information that would be useful to have.

Also reworked the loader so that it delegates to the various
format-specific loaders to answer questions of what file to actually
load, etc.  (For example, a user may select the ENVI .hdr file, or
the GeoTIFF .tfw file, and WISER needs to be smart enough to find the
actual raster data to open, based on the user-selected file.)  I was
pleased that the refactoring was relatively small in scope to get the
GeoTIFF support added - but, still a lot of work to do on it.