Ability to "lock" spectrum-plot to specific datasets
Previously, the spectrum-plot window would switch to whatever dataset
was clicked on in the main or zoom windows.  This was not helpful when
using a mineral map to navigate around a core sample.  The spectrum-plot
window now includes a dataset-chooser that lets the user "lock" the
spectrum-plot window to only show spectra from a specific dataset.
Of course it also includes the ability to switch back to "show spectrum
from clicked dataset" if the user so desires.