Basic functionality for displaying geographic coordinates
The WISER status bar can now show the geographic coordinates of clicked
pixels.  The UI can handle datasets with no coordinate reference system
(CRS), or with just a geographic CRS, or a projected CRS (which also
contains a geographic CRS).  If a CRS has "degree" as the units, the UI
is able to switch between degrees/minutes/seconds and just degrees.

The "go to coordinate" functionality is only beginning to be done, but
the display work is pretty far along.

Still a lot of refining to do, but it's a pretty good step forward.