Bugfixes and tweaks to geo-coordinates UI
Fixed a small bug with coordinate-transformations, where the GDAL code
was returning a tuple of coordinates and then WISER was trying to
mutate it.

Updated the "Goto-coordinate" UI to have "Units:" as a separate label.

Improved the precision shown in the UI, since it was 15-digit full-on
double precision before.