Goto-coordinate transformation code is mostly complete
A few other tweaks as well, but the main content of this commit is that
the Geographic Coordinates dialog is now able to transform geographic
coordinates in some CRS back into pixel coordinates.  There is still a
lot of debug output at this point, but since it works and it was a bit
tricky, it seemed good to get it checked in.

Also improved the UI behavior a bit, disabling and enabling widgets
based on whether the CRS was geographic or not.

Also did a bit of code refactoring.