Geographic coordinates "goto" works; usability improvements
The "Goto Coordinate" button now works in the geographic coordinates
dialog.  The implementation should be able to handle degrees,
degrees/minutes/seconds, northing/easting, as well as direction
indicators at the end of these inputs.

The "image-coordinates" widget was also reworked to be
visible/enabled/disabled based on some dataset updates.  The
geographic-coordinates toolbutton is always visible now, but is
merely disabled when it is not relevant.