Add Windows WISER icon.  Bump version to 1.1a7-dev0.
The Windows platform doesn't automatically use the file's icon as the
running application's icon.  To fix this, the wiser.ico file was added
to the WISER resource-bundle, and applied in the main application at
startup.

The Windows installer and pyinstaller invocation were also both
updated to use the wiser.ico icon.