Spectral-plot "load spectra" button can also import text spectra
It was confusing that the "load spectral library" button in the
spectral-plot window could only load .sli files, but you had to
go to the application's "File" menu to import text spectra.

The UI interaction is updated so that the "load spectra" button now
pops up a menu that lets the user choose to load a spectral library,
or to import ASCII spectra.