Update Windows build process to use version.py directly
The Windows NSIS .nsi file now looks for a WISER_VERSION definition
instead of having the version hard-coded into the .nsi file.  The .nsi
file also adds more registry-values for the uninstaller now, so that
the current version apperas in the "Apps" control-panel display, and
the WISER icon is also shown.