Tentative fix for bug #15
The spectrum-plot code wasn't respecting the units being reported for
the spectrum being displayed.  The function that generated the plot
would accept the units to use, but it was defaulting to nm.  The fix
was to actually pass in the unit value, like it should have been doing.

This fix is tentative because it really requires more testing to verify
that we don't have any other misbehaviors with mixed unit types, and
also that we didn't introduce any other regressions.