Better handling of small screens with contrast-stretch config UI
(Issue #16)

The contrast-stretch config UI was taller than some smaller screens
would be able to show.  A couple of tactics were employed to address
this issue.

The widget that shows the configuration for each color channel is
updated to handle resizing more elegantly.  More work can be done on
this; the 256-bin histogram becomes rather apparent when the
channel-widget config is widened substantially.

A scroll-area was worked into the stretch-config UI to hold the
color-channel widgets.  When the dialog is shown, it attempts to resize
itself to a size that allows all channels to be visible, unless that
would make the dialog taller than the screen.  If the screen is small,
the dialog will stay within the screen, and the scrollbar will show.

There is still a bit of a sizing issue on macOS; the contents of the
scroll-area are slightly narrower than the other dialog contents.  At
some point we can address this, but not now.