Propagate spatial and spectral metadata through band-math expressions
This commit contains a number of changes to support the propagation of
spatial and spectral metadata through band-math expressions.  The
approach chosen was to use the expression-analysis phase to identify
where spatial and/or spectral metadata should come from, for every
single term of a band-math expression.  This information is propagated
through analysis, so that after evaluation, WISER can simply apply the
appropriate spatial and spectral metadata to the result object.

There may still be some bugs in this code - every operator had to
change, and a number of other changes had to be made to the base
data-types to support this feature.  Some basic testing has been done,
but bashing on this functionality is likely a good idea.