Fix bug #31, band-math expressions like "1 - b" don't work
This simply wasn't supported by the initial version, since I didn't
think about expressions like that.

Subtraction is now implemented like addition, where A+B may be
transformed into B+A to simplify the logic.  Of course, with
subtraction, it's A+(-B) that may be transformed into (-B)+A.

This commit also includes a refactoring of some of the helper operations
for implementing band-math calculations, moving them into the top-level
package.