Changed view port highlight logic to only show
display highlight in the rasterview's that
have the same dataset as the zoom pane. If its
linked then all rasterview's that are
geographically compatible will show the yellow
highlight box.

For context pane, this MR makes only one
list of highlights display. This list
corresponds to the rasterview's that
match the dataset ID of the displayed
context pane.