Fixed problem in geo ref with selected points
disappearing after the _georeference function
was called. It was because we would sync the
table with the entry list but use the pixel
value instead of the spatial value. When we
add in new points, the signal cellChanged is
triggered with extracts the point value
from the table and makes the reference gcp
equal that point value. But that point value
is assumed to be a georeferenced value.