Fixed issue with other coordinate systems causing
WISER to freeze. The problem was due to an infinite
function calls between some Qt signals and slots
in the scroll area of the Q table widget. The
coordinate system that was causing this took
just long enough to calculate most of the time that
it put WISER into that state.