Basic a+b in subprocess, Progress #1 (#170)
Description
This PR contains the refactor that lets us use subprocessing for bandmath operations which lets us not hang or slow down the GUI.

How it works
To do these I created a way to serialize the data for the classes RasterBand, Spectrum, and RasterDataSet. I also had to make spatial_metadata_source into the class SpatialMetadataSource so it could be pickleable. I did the same for spectral_metadata_source (made into SpectralMetadataSource class).

I created the class ProcessManager and ParallelTaskProcess to handle the asynchronous execution. I may refactor these in the future.

The way it currently works (this is still in development), is that eval_bandmath_expr creates a ProcessManager class. We pass in a function that we want to run in a subprocess into the ProcessManager instantiation. We make the arguments be in a serialized format before we pass them into ProcessManager with the function. The ProcessManager class contains a ParallelTaskProcess class that has slots. We connect the succeeded, progress, and error slots up then start the task. When the task is finished, the results are put on a queue that the main process reads from.

Why its needed
We want this so we can do batch processing and so the UI remains responsive.

Tests
I changed the tests for bandmath to work with the subprocessing. I also added tests for the serialization of RasterDataSets, RasterBand, and Spectra.