Adding faster environment load with micromamba (#228)
Adding a better way to load our dev environment with micromamba and using our conda-lock files. The micromamba environment should recache itself everytime our lock file changes. It makes a linux cache.

The tests now run in the micromamba environment. (The code that micromamba uses to start up the environment is in the login scripts of a shell which is why we do "shell: bash -leo pipefail {0}").

There was an issue with the tests not working on linux because linux could use the GetVirtualMem functions that wiser has. The issue arose because we deleted the underlying dataset which caused the memory mapped array to be invalidated. The fix I implemented was to not use GetVirtualMem until we have more time to investigate how to incorporate it while also having multithreading. This problem is better describe in issue #227 .

Also made a context for multiprocessing whenever we create a multiprocessing object to ensure that we are using the start method.