Adding initial sphinx docs for pre-open source documentation #231
Adding new book themed wiser docs. They are currently incomplete, but I will populate them as I prepare for open source.

The point of them is for me to add documentation on how to use WISER both as a user and as a developer. Currently I just have it set up so that I can add more documentation in the future.

Just run make html (or any of the commands sphinx lets you do after make) to build the html documents when you are in the folder doc/sphinx-general-wiser-docs