Adds docs on how to set up the users dev and prod environment (#233)
As the title says, adds documentation on how the developer should set up their dev and prod environment to 'General Wiser Docs'.

It also adds docs on how the developer can change the lockfile if needed.

A makefile is added to make these commands easy for the user to run and make commands dependent on earlier steps.

This PR also deletes the dockerfile, entry_point.sh, and wiser-source.yml files which were all used to create the docker container that we no longer use.