Add ruff linting and formatting to WISER (#238)
As the title says. But I will have to decide what rules to ignore and to keep. Reasons for ignoring rules will be below. I added ruff's linting and formatting as a pre-commit hook and added it as a job to run before the tests. I also updated the dev-conda-lock.yml to include the pip library pre-commit which helps us make pre-commit hooks. Updating dev-conda-lock.yml is why this commit is so big.

Ignoring F403: We use wild card imports for PySide2 objects. It is very useful.

Ignoring F405: We use wild card imports for PySide2 objects. It is very useful.

Ignoring E722: We use bare except in multiple locations in WISER already. It would be good to fix this, but currently we don't have the bandwidth. Look at Issue #239 to fix.

Ignoring F401: Unused imports are used in some places in WISER to provide sys path modifications and to import a collection of things. This means we have to carefully go through each unused important to make sure getting rid of it won't ruin anything by neglecting a sys path modification. Again, we don't have the bandwidth right now, but would ideally want to fix in the future. Issue #240 is made to address this.

Other

Currently line-length in pyproject.toml is set to 110. We would ideally want to lower this to 88. WISER has just gone too long without a linter so there are hundreds of long lines to shorten. Line length shortening isn't a safe fix which is why this is a manual process. Issue Shorten line-length in pyproject.toml to 88 #241 is made to address this

* Added pyproject.toml file for ruff

* Runned ruff check --fix to fix safe fixes

* Changed == None to 'is None'

* Changed == True and == False to be more concise. Commented out some unused variables

* Going through ruff fixes

* Got rid of many unused variablse

* Fixing ruff linter issues

* Deleted unused variable

* Turned lambdas into functions

* Removed True== and False==

* Making ruff check fixes

* Ruff check passes

* Changed pyproject.toml for ruff formatting. Added generated _ui.py to exclude and set maagic-trailing-comma to be respected

* Formatted code with ruff

* Formatted the rest of WISER. Passes 'ruff check' and 'ruff format'

* Added pre-commit dependency to dev yaml

* Added pre-commit to dev environment and refresh .yml

* Added pre-commit yaml file that runs ruff formatter and linter

* Added pre-commit hook using .pre-commit-config.yaml.
It should run just the same as running ruff locally

* Updated yml tests including linting before the tests start

* Renamed dev-tests to dev-CI since we having linting which makes this more of a Continuous Integration (CI) workflow

* Changed line-length max to 110 instead of disabling E501.
Changed a lot of files to ensure no line lengths were above 110. Reformatted as well.