Stop duplicate Github actions when pushing to a pull request for all branches (#243)
We want to get rid of duplicate actions. We use fkirc/skip-duplicate-actions@v5 to do this (community github action that uses the Check Runs API, learn more here).

We also only want to automatically run the linting and tests when we are pushing to and we are PR targeting an important branch (like main or a release branch (rel/**)).
I did this by only running the Run tests and linting with linux-cached-dev workflow when a push is made directly to main or a release branch (re/**) or when a push is made to a PR that targets main or rel/** is done.

Currently, the only possible way duplication is to have a PR that goes from main to rel/** or rel/** to rel/** or rel** to main.

These PR implements these wants.

* Added code to remove duplicate actions for when we push owards a pull request.

Also added code to only automatically run linting and tests on branches matching:
main
rel/**
feat/**

* Edited on: push: and on: pull_request: in dev-CI.yml to only do those for main and rel/** branches

* Added name to the pre check commit that would
skip duplicate actions