Update .md files in doc folder (#237)
Update the .md files in doc folder to be current. The .md files like devenv-mac.md and devenv-win.md are outdated. The other .md files that I made changes to in this PR were also outdated.

* Updated codesign-mac.md documentation

* Added codesign-mac.md documentation to general-wiser-docs sphinx pages

* Updated devenv-mac documentation to use the make install-dev-env
command to build the environment instead of installing packages one at a time as this is brittle

* Updated windows documentation to be current

* Redid section headers to have a better layout in the sphinx docs and also to make more sense.

Also added a section to environment-setup.md about how to actually run WISER

* Added a few changes to plugins.md