Add current continuous integration setup to docs (#247)
As the title says. We explain how the current continuous integration setup works.

I explain how the pre-commit linting works and how the environment set up for testing works.
I explain when the github actions linting and testing happens

Commits:

* Added documentation on linting/formatting and added pre-commit installation to install-dev-env in etc/Makefile

* Added continuous-integration docs

* Added continuous-integration.md to toctree for the developer section.

Also increased the max depth of the navbar so it would show (before it was at depth 4 by default)

* Added that testing happens after linting