Add a test mode CLI argument to WISER (#250)
We want to add a test mode CLI to WISER so that we can run tests in the production version of the app. The overall goal of this is to be able to automatically run tests on each of the build artifacts for each of the platforms.

This PR adds in a test mode to WISER's CLI and tests one file. It also changes the lock files to put pytest into the prod lock file because we will need it to test out production environment.

I did test to make sure the built artifact could run the test mode on windows and macos

Closes #252

* Added a command line argument to run WISER in testing mode.

This is meant to be used to test how our app works in our production environment
(after pyinstaller builds it into an executable), so I also added pytest to our
prod environment dependencies. I will rebuild on the next commit

* Updated production environment to have pytest in it.

Recreated dev and prod lock files to reflect this

* Found bugs that didn't let us run ./WISER.exe --test_mode

and fixed. First bug was that auto loaded pytest plugins weren't being loaded.

Second bug was that we weren't packaging the tests and test_datasets into
final executable so tests wouldn't run

* Fixed quick WISER-macOs.spec error in gdal plugin file path

for jp2 plugin

* Fixed quick WISER-macOs.spec error in gdal plugin file path

for jp2 plugin having extra Library folder

* Fixed macos jp2 gdal plugin dylib having wrong extension