Fixed issue with continuum removal on branch 1.3b1 (PR #267). Merging into main (#268)
Continuum removal was broken. Fixing issue with it here.

First there was an issue with continuum removal on an image returning a different value than continuum removal on a spectrum. This was because continuum removal on an image ignored the bad bands (and the bad bands had noisy data that messed up the hull).

The second issue was that the UI files for continuum removal weren't packaged with the application. These should both be fixed.

Closes #265

* Fixed Ui issue with continuum removal but found other issue

* Added test for continuum remove on single spectrum

* Made tests for comparing continuum removed image for a many band image and
compare output to correct continuum removed single spectrum

* Tests finally passed for continuum removal. Should work correctly.

Code still needs to be cleaned up

* Got rid of comments

* Cleaned up imports

* Added a test for subsetting and continuum removing at the same time