Fixed issue with conda environments not getting found if they weren't
the one that I was in. Issue stemmed because my environment
was saying it was the root_prefix (which would normally) default
to \anaconda3\ so then you can search the environments from
there but it was defaulting to \anaconda3\envs\wiser-dev which
messed everything up. I made a fix that adds in the list of all
envs so even if this mistake were to happen it would still find the
libraries