Make It Easier to Add Plugins To WISER (#264)
This PR makes it easier to add plugins to WISER by putting new GUI elements to do this.

Adresses #291

* Edited app_config.ui file to add basic necessities

for easily adding plugins. I added the necessary
UI pieces.

* Added code to help get part of the fully qualified path

and get the base directory for that fully qualified path

* Added the rest of code to let us select a file

and extract all the information that WISER needs
to get the plugin from that file

* Made adding from environment manager work for conda

* Added doc string to _load_plugin_from_file. Made plugin class name ledit

not editable. Remove bug with setting the env name when adding
a plugin class.

* Changed plugin select file push button to say

'Select File'

* Finished editing the UI to make the new way to add plugins more intuitive

* Added message box to tell user if
the environment could or could not be found

* Added tests for adding plugins by file path

* Made settings starting tab be the general tab

* Fixed multiple of the same conda environment being able to be added

* Fixed issue with conda environments not getting found if they weren't

the one that I was in. Issue stemmed because my environment
was saying it was the root_prefix (which would normally) default
to \anaconda3\ so then you can search the environments from
there but it was defaulting to \anaconda3\envs\wiser-dev which
messed everything up. I made a fix that adds in the list of all
envs so even if this mistake were to happen it would still find the
libraries

* Fixed issue with removing base_dir_abs path